requirements.txt
python-dotenv==1.1.0
requests==2.32.4
ruff==0.11.13
urllib3==2.4.0
